Spike count record

+++ ml_genn/ml_genn/losses/mean_square_error.py
@@ -10,8 +10,9 @@ class MeanSquareError(Loss):
         # Add extra global parameter to store Y* throughout example
         flat_shape = np.prod(shape)
         egp_size = (example_timesteps * batch_size * flat_shape)
-        model.add_egp("YTrue", "scalar*", np.empty(egp_size))
-    
+        model.add_egp("YTrue", "scalar*",
+                      np.empty(egp_size, dtype=np.float32))
+
         # Add sim-code to read out correct yTrue value 
         model.append_sim_code(
             f"""
@@ -29,10 +30,10 @@ class MeanSquareError(Loss):
         if y_true.shape != expected_shape:
             raise RuntimeError(f"Shape of target data for MeanSquareError "
                                f"loss should be {expected_shape}")
-        
+
         # Copy flattened y_true into view
         genn_pop.extra_global_params["YTrue"].view[:] = y_true.flatten()
-        
+
         # Push YTrue to device
         genn_pop.push_extra_global_param_to_device("YTrue")
